Changed the user points attribute to be decimal instead of an integer

+++ api/models/RecycleUser.js
@@ -42,6 +42,6 @@ module.exports = {
   },
 
   afterCreate: (newlyRecords, next) => {
-    User.addPoints({ userId: newlyRecords.user, points: parseInt(newlyRecords.scrapKg)}, next)
+    User.addPoints({ userId: newlyRecords.user, points: parseFloat(newlyRecords.scrapKg)}, next)
   }
 };
